Rename q and schedule back to queue and task
