+++ utils/task.lua
@@ -4,7 +4,7 @@
 -- github: https://github.com/Valansch/RedMew
 -- ======================================================= --
 
-local Queue = require 'utils.q'
+local Queue = require 'utils.queue'
 local PriorityQueue = require 'utils.priority_queue'
 local Event = require 'utils.event'
 local Token = require 'utils.token'
